共通処理をまとめるベースクラスを作成
